+++ src/main/java/types/TSPdata.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 
 public interface TSPdata extends Serializable {
     void save(String fileName);
-    double cost(int... path);
+    long cost(int... path);
     //int[] kRandom(int k);
 }
